app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
neo4j==5.14.1
pandas==2.1.3